Grid3D: Moving model definition to the test directory
