+++ opesci/__init__.py
@@ -1,4 +1,3 @@
 from variable import *
 from fields import *
 from staggeredgrid import *
-import grid_test
